change last_update_ in EventLogRebuildingSet only if the update was successful
Summary: This diff fixes a bug where last_update_ was updated irrespective of whether the delta successfully applied to the EventLogRebuildingSet and thus causing the version_ maintained within the replicated state machine to diverge from that of EventLogRebuildingSet. This patch fixes it by updating last_update_ only if delta was applied successfully. And it does it for all kinds of event log delta (Note that previously, SHARD_UNDRAIN and SHARD_DONOR_PROGRESS will not result in updation of last_update_). There is no side effect of this since we do not use version to check equality of two ShardAuthoritativeStatusMaps and hence should not result in additional updates to workers or SHARD_STATUS_UPDATE messages.

Reviewed By: davidvigier

Differential Revision: D9776389

fbshipit-source-id: 4c4f381fcf233a01aeea2d482160d948d2035148