Decouple SocketType (DATA/GOSSIP) from ConnectionType (SSL/PLAIN)
Summary:
The intent of this change is to tighten security by making sure everything is encrypted for cross ssl-boundary traffic, including gossips. It requires disabling the send-to-gossip-port option.

Because of the way the Sender code is implemented, it is currently not possible to use SSL for a gossip socket. The supported socket types are DATA, SSL and GOSSIP, which makes SSL and GOSSIP incompatible (or impossible to set simultaneously).

This diff changes the Socket and Sender API to use the SocketType (DATA/GOSSIP) and the new ConnectionType (SSL/PLAIN) enums.

This allows sending gossips over an encrypted connection. Now, if the gossip connections cross the ssl-boundary, and they would normally go through the data port, they will go to the SSL port instead. Note that the dedicated gossip port stays unencrypted as it would be a backward incompatible change otherwise.

Eventually we should get rid of SocketType altogether. Socket should not need to know what kind of messages are going through. And eventually we should also get rid of the gossip and ssl ports, and rather have a single listener that can assign the socket to the proper worker thread. But that is beyond the scope of this diff.

Reviewed By: scsiguy

Differential Revision: D9784485

fbshipit-source-id: 244fc7e1f4a9dabdbef5241dc7e35c9ad12fd4ff