StorageMembership: state definitions and transitions
Summary:
Storage membership will be the dynamic part of the storage
configuration. (i.e., StorageConfiguration = Storage membership +
Service discovery info).  This patch defines the basic states, allowed
state transitions and conditions for state transitions of storage
membership in LogDevice.

Note that the **state machine** that manages the state transitions,
along with the API implementation that manipulates the membership is
not included. This patch only contains the state definition and
transition rules.

Reviewed By: shixiao

Differential Revision: D9414308

fbshipit-source-id: bc2e8d6d411af45eced89229ffc95a6f72906bcf