StorageMembership: add a new provision transition
Summary:
Add a new type of PROVSION transitions which are used to
create the initial storage configuration for a newly created cluster.
In such case, all provisioned shards will be directly transitioned into
READ_WRTIE state instead of NONE. This will be convenient for the cluster
turn up process.

Note that for safety the provsision transitions are conditioned on
EMPTY_VERSION (0), so that it can only be used at the very beginning of cluster creation.
I have repurposed the INVALID_VERSION to be EMPTY_VERSION, so that EMPTY_VERSION
becomes valid but always correspond to an empty membership.

Depends on D9414308.

Reviewed By: shixiao

Differential Revision: D9598251

fbshipit-source-id: 816bff25930ce07bf607a22f4fbc62c93922fc46