Using a shared SSL Context and cleaning temp buffers after done
Summary:
Some smaller fixes in Logdevice to reduce memory consumption
a) reuse an sslcontext. We currently even create one for non ssl connections and then ignore it which is a bit ugly logically but nicer code. Willing to do the checks outside as well
b) Clear the buffer objects after the final response has been created

Reviewed By: davidvigier

Differential Revision: D9882020

fbshipit-source-id: 788a441a825c5bf72623faae5fc7c76bd5199297