Reclaim unused memory in AdminCommandClient every few calls to readDataAvailable
Summary:
The AdminCommandClient allocates 8KB buffers for every call of getReadBuffer(). However the socket may not use all of it. This wasted memory is only released in the end when it is done with the connection.
This diff merges the buffers every few calls to readDataAvailable. This releases previously allocated buffers and reduces the memory waste.

Differential Revision: D9885695

fbshipit-source-id: 7280a3f7ed7f6e20bf3f989e95fbe0bc6e47d23f