Deregister read callback from socket upon destruction of AdminCommandConnection
Summary: When the AdminCommandConnection is destroyed, we call AsyncSocket::close() which in turn executes the readEOF handler. This causes connection to mistakenly execute the callback and be excluded when computing how many nodes timed out. To fix this bug, I deregister the read callback from the socket before closing it.

Reviewed By: adrienconrath

Differential Revision: D9932723

fbshipit-source-id: b3aba5b2edfe44a7380419cbec602285460e5795