Use a temporary evbuffer to gather admin command output
Summary:
When using SSL on the admin command port, the server instantiates a bufferevent_openssl_filter to wrap the bufferevent that owns the socket. The way it works is every time the output evbuffer of the bufferevent_openssl_filter is updated (some data is written to it), it immediately calls into openssl to write a TLS record containing the appended data into the underlying evbuffer. This is very inefficient in the context of admin commands because each line of the output is individually appended to the evbuffer. So basically every line is encapsulated in its own TLS record. So instead of fully utilizing the 16kb TLS record size, we only use a few bytes.
In particular for commands such as "stats" that have hundreds of lines (of just a few bytes each), this translates to a lot of records, which takes longer to produce and costs more resources.
This also affects the client side (AdminCommandClient) which ends up needing to allocate more read buffers (wasting memory) and AsyncSSLSocket calling hundreds of time the read callback (wasting CPU).

In order to address that issue, I now create a temporary evbuffer which I pass to the command class and which accumulates the output. Then I give ownership of that buffer to the output buffer in one call, which allows the bufferevent_openssl_filter to break it up into larger TLS records.

Reviewed By: adrienconrath

Differential Revision: D9933382

fbshipit-source-id: dd73d994070a89242706865a5a1b2f163a105ce9