Fixed NodeSetTest.DeferReleasesUntilMetaDataRead test flakiness
Summary: This test is flakey, because occasionally the sequencer node is terminated before it sends releases for the last LSN of the first epoch, which causes it to be unreadable until recovery completes. Since the test blocks recovery from completing, it hangs forever. So the fix is to read the first epoch to ensure that releases have been propagated before proceeding with terminating the sequencer.

Reviewed By: mcrnic

Differential Revision: D9972115

fbshipit-source-id: 7a8e580a89a50b83f6dd4a466fa7c1bea883c544