Require node "roles" config attribute (daemon, client, and utilities)
Summary:
The node roles attribute, is an array listing the services provided
by a given node in a LogDevice cluster. The currently supported roles
are "sequencer" and "storage". The roles attribute reflects the intended
steady-state operation of a node and is set at the time a node is
provisioned for inclusion in a cluster.  It does not change while a node
is a member of a cluster. However, there are times when a node can't serve
its provisioned role(s). For example, the node may have failed or degraded
storage, or is suffering from some other hardware, network, or software
induced issue. In these situations, roles can be temporarily disabled by
changing other attributes:

- Sequencing Role: "sequencer" : false
- Storage Role:    "storage"   : "disabled"

All role specific attributes are gated by the presence of the respective
provisioned role. For example, the "sequencer" attribute is ignored unless
a node also has "sequencer" as an element in the roles attribute.

The motivation for this change is to simplify operations with heterogeneous
nodes that serve different roles. The "roles" attribute represent the long
term intended state for nodes and, since they are persistent, can be used
to provide properly filtered views of the cluster. (e.g. How much sequencer
capacity is disabled, and how much of that is for nodes currently in repair).

Reviewed By: jpdborgna

Differential Revision: D7743786

fbshipit-source-id: 03eaccbeaab7bba83a57ced2e0eaeea495f154a6