track size of trash files
Summary:
When ratelimiting sst file deletions, we have to be careful so the trash
doesn't grow too large. This diff adds a metric so that we can track this. To
let us track this per shard, we switch to using one SstFileManager per shard.

This also makes the ratelimit setting specify ratelimit on a per-shard basis.
This is fine since it currenlty isn't used anywhere.

Reviewed By: scsiguy

Differential Revision: D8876616

fbshipit-source-id: 0bd23a51945c2902ef49fa24922c55ea8b549d11