Prevent deadlocking in fatal signal handler while unmapping RocksDB caches
Summary:
We unmap the RocksDB caches in our fatal signal handler to reduce the size of our coredumps.

When the process gets a SIGSEGV (or SIGABRT and few other friends, see `server/main.cpp`), it could have the lock to the RocksDB cache, when the segfault happens to be in that same thread, trying to call `ApplyToAllCacheEntries` with `thread_safe=true` will deadlock.

This will leave the process in a zombie state.

Would like some feedback whether removing `thread_safe` here is the correct approach.

Alternatively I'd not even bother with unmapping memory for `SIGSEGV/SIGBUS/SIGILL/SIGFPE` given the memory / process state is unpredictable.

Reviewed By: yiwu-arbug, al13n321

Differential Revision: D9849895

fbshipit-source-id: a3d00d4a3011bb46e390def0b0e71ff12ca3f78c