Clean up metadata log configuration + fix RemoteLogsConfig for
Summary:
After D6940579, RemoteLogsConfig uses LOGS_CONFIG_API_Message. The response for
this is based on LogsConfigTree on the server, but this does not include the
internal logs (stored separately). Therefore, RemoteLogsConfig returns that the
log does not exist. This breaks e.g. TRIM operations with on-demand-logs-config
enabled.

This diff:
- Cleans up passing around metadata logsconfig callbacks and handles this from
a single source: Configuration.cpp
- Handles the internal logs cases from Configuration.cpp as well

Passing around the callback was tech-debt created a long time ago, but can now
be abstracted away by using the calls to Configuration instead.

Reviewed By: tau0

Differential Revision: D10008506

fbshipit-source-id: 2bba4b8fecd6b69d6c960af790cd314cd89c0640