Move STORE::onSent() and STORED::onReceived() into server/ to give them access to ServerWorker
Summary:
I'm going to need this to dispatch STORE replies to ChunkRebuilding-s during rebuilding partition by partition. That's because, unlike the tech-debty LogRebuildings, I'm putting the registry of ChunkRebuildings in ServerWorker rather than Worker.

This diff only moves the routing of stores with REBUILDING flag set, which is almost certainly going to be server-specific forever. Leaving routing to Appender and Recovery in both client and server in case we ever want to run those on clients. Note that even if I moved Appender and Recovery stuff to the server-only part, I can't make the message server-only because it's also used by meta-fixer.

The implementation ended up a little awkward because this is different from all other messages in that it's used by both client and server, but the logic for routing the reply is different.

Reviewed By: sh00s

Differential Revision: D9819879

fbshipit-source-id: 2e43e9a5d21489ba4407e98fb90c2d320277d1de