Use all send all timer even when SCD greylisting is enabled
Summary:
We still suffer from some edge cases in the SCD greylisting feature that can rarely cause a reader to stall indefinitely
(more details in T34193119).

In this diff, I am re-enabling the all send all timer even when SCD greylisting is used so that we at least have a
guarantee that the reader will eventually rewind.

This diff also removes a condition for when to actually do the all send all timer as I don't think it's necessary to
have the logic we need.

Reviewed By: lucaspmelo

Differential Revision: D9967220

fbshipit-source-id: fcb48e43ee7a124925c944ddc6d13c4d71156cd0