Set sequencer weight to 0 if sequencer=false in config regardless of the value of sequencer_weight
Summary:
Do not read the sequencer_weight attribute if sequencer=false in the config and set the weight to 0.
Previously the code would consider a sequencer enabled if sequencer_weight > 0 even if sequencer was set to false in the config. This would cause newer clients and older servers to disagree on where a sequencer should run.

Differential Revision: D10133633

fbshipit-source-id: 618e29e2fa0bf0cab5ebe43211651a63ef5f6cd0