Bump default grace period to fail-safe 5s
Summary:
The default grace period is currently 100ms, which can be too low and cause
tests to flap unless they specify an appropriate grace period themselves. This
diff changes the default to 5s, which should be more than enough. Customers
that prioritize low latency should adjust this, though it's also capped to the
client timeout, which they probably already set.

Differential Revision: D10202253

fbshipit-source-id: 1c0e989abe25ffe9886e92d04cc50e6ed8e4579b