Fix bug when draining items after transitioning flow group from enabled to disabled.
Summary:
After traffic shaping is disabled, if time exceeds when draining a flow group
queue there was bug in logic which skipped draining the remaining items. This
diff fixes the bug.

The FlowGroup::run used to check if priorityq is empty and PRIORITY_QUEUE class
bucket is empty to make a decision about time limit execeeded. This works when
the flow group is enabled and is an optimization to avoid an extra run call. But
when the flow group is disabled PRIORITY_QUEUE class bucket is empty and if
time limit was actually exceeded during drain run method incorrectly returned
false. And items in the queue were never drained, and new items kept on getting
added to queue.

This bug though hit in read path could have happened for any traffic class.

Reviewed By: scsiguy

Differential Revision: D10174258

fbshipit-source-id: 0de9d9b704629cd314f8067d6adaad6623be256d