ClientReadersFlowTracer: adjust time_lagging computation for redelivery timer
Summary:
- Did some refactoring in `ClientReadersFlowTracer` so the ideas could be more easily implementable;
 - Added callback hooks to adjust for time lag accumulated when redelivery timer is active.

Basically we will ignore all samples that are gathered when redelivery timer is
active and we will normalize the remaining time samples so we don't have to
keep track of the timestamp in which we gather those (i.e. we will round their
timestamps to the beginning of the 1 min window).

This should be fine as most of our samples should come within <100ms from the
beginning of the 1min window.

Reviewed By: al13n321

Differential Revision: D9483111

fbshipit-source-id: 73c0f19548c89b2cf8241c6a142720ef484b07ee