Reject rebuilding stores above the last_released_lsn and LCE and kick off purging
Summary:
This patch fixes a case where a node that does not have an updated LCE or last_released_lsn can accept a rebuilding store which is later deleted by purging eventually when the node receives a release for the log. The fix is to reject the store with E::DISABLED (rebuilding will retry the same node due to deterministic copyset selection and treats this as transient failure) and kick off purging by treating the store as release message. This is safe because rebuilding is run only after recovery is complete , so anything written by rebuilding is guaranteed to have been released by sequencer.

Why isn't it sufficient to look at local LCE alone? Because the rebuilding could send a store for a record in current lsn as well (since we get until_lsn which is the next_lsn from sequencer and wait till next_lsn is released before proceeding to rebuild). In that case, rejecting store by checking against LCE alone isn't sufficient since LCE will never move to current epoch

Reviewed By: runemaster

Differential Revision: D10117484

fbshipit-source-id: 7af6e10bea08218008c8ec8b5c68d15c59f046e7