Replaced LibeventTimer with Timer interface which doesn't have libevent in
Summary:
This commit is part of a preparation for redesigning workers.

One thing which complicates this redesing is `libevent` exposure to the code. Each time we create LibeventTimer we need to pass a `struct event_base* ` which will not be available after redesign, because in the end all timers should be inside one of thread pools and our Timer API shouldn't know anything about libevent.

What has been done:
1) Introduced new Timer.h which is a wrapper around LibeventTimer which don't have any libevent types in signatures.
This will allow us in the future to change Timer implementation for the new one, which probably will send requests to create timers to one of the thread pools.

2) Changed all of logdeviced code from LibeventTimer to Timer class. And some of other code. Ldbench code hasn't been changed , because it has a bit different workers which will not be redesigned.

Reviewed By: mcrnic, sh00s

Differential Revision: D10052805

fbshipit-source-id: eacf60e8a209c2a589541c8531a5f1be885ed327