LogsDB: interleave partial compactions with retention compactions
Summary:
Special effect that caused T34179816: logdeviced starts after a day of being stopped; it has a long backlog of retention compactions to go through, which takes a few hours during which no partial compactions happen; this node is a recipient for a rebuilding; soon it accumulates 50 pending partial compactions and stalls all rebuilding writes for a few hours, stalling the whole rebuilding.

This diff is a simple tweak to compaction prioritization that makes sure partial compactions don't get completely starved.

Reviewed By: sh00s

Differential Revision: D10002305

fbshipit-source-id: 8f1a6068aff9caa8b3f80ee90a52aa3412b3272d