Split ZookeeperConfig into separate UpdateableConfig
Summary:
This diff extracts `ZookeeperConfig` out of the main `ServerConfig` file.
It is the first step of taking apart the `ServerConfig` such that each of them can be updated and subscribed to independently.

This means we can start using service discovery components in the future responsible for updating the `ZookeeperConfig`.
I'll be gradually refactoring / taking apart `TextConfigUpdater` to allow for a more flexible and consistent update workflow (including `LogsConfigManager` and `RemoteLogsConfig`).

As part of the node membership project, `NodeConfig` is next in line.

This diff also removes `ZookeeperConfig` config syncing (read side). This section is now ignored, but still embedded in the `CONFIG_CHANGED` message in order to provide backwards support for older servers. Clients already don't need this section. Once all servers are updated, the extra logic there can be removed.

Reviewed By: AhmedSoliman

Differential Revision: D10163066

fbshipit-source-id: 74cd5633c8ad148d2f48de52a7ed0f2eedb73b43