State machine for re-replicating a chunk of records
Summary:
ChunkRebuilding is like RecordRebuilding but for a bunch of records at once. No support for rebuilding without WAL yet. For now it's implemented as just a wrapper around a bunch of RecordRebuilding-s; in future I'm going to make RecordRebuilding{Base,Store,Amend} to operate on a chunk instead, and maybe even send all records in a single STORE message.

This diff implements ChunkRebuilding, sets up routing of STORE onSent() and STORED onReceived() to it, and adds a stub of ShardRebuildingV2.

There's some funky stuff with boilerplate PayloadHolder-s inside ChunkData, and aliasing shared_ptr to them, all to be able to pass shared ownership of the chunk to STORE_Message. I thought long and hard how to avoid it and couldn't come up with anything better.

Reviewed By: pvjykumar

Differential Revision: D10096961

fbshipit-source-id: 0a909f122d1b0488067f84648e904cf36515d263