BuildInfo plugin
Summary:
Separates the `BuildInfo` plugin from the `LegacyPluginPack`.

Since `BuildInfo` is loaded via `Processor`, I had to pipe through `PluginRegistry` to `Processor`. Since `Processor` isn't aware of the differences between the `ClientPluginPack` and `ServerPluginPack` and accepts either as the plugin argument (but in the new framework they are different plugin types), I had to leave `plugin` as an argument to its constructor as well for now. `ConfigInit` may create a `Processor` indirectly, so passing both in there as well. In the future this will be removed and only `PluginRegistry` will be passed to `Processor` and `ConfigInit` constructors.

Reviewed By: calebmarchent

Differential Revision: D10405506

fbshipit-source-id: 04c4eac8586df3386f4e909f739fd77e442f1fc1