Move ZK rc to LD status function to ZookeeperClientBase
Summary:
While ZookeeperClient returns the ZK ret code so that no information is lost,
it's convenient to provide a way to interpret ZK return codes as LD statuses.

The error codes returned remain the same in all but the following cases:
* for unexpected ZK error code, return `E::UNKNOWN`;
* for `ZBADVERSION`, return `E::VERSION_MISMATCH` instead of `E::AGAIN`, which
  is useful in places like NodesConfiguration; (I audited the callsites and
  ensured no special handling was done for `E::AGAIN`.)

This diff also changed an ld_check (before logging) to ld_assert.

Reviewed By: sh00s

Differential Revision: D10347844

fbshipit-source-id: 31ff70d34d02fce087f85804950c077b71c18fd0