Support filtering based on directory in logsdb
Summary:
Usually a rebuilding donor needs to read only a small percentage of the logs+epochs it has, because most logs have nodeset not intersecting rebuilding set. When records are big and CSI is effective, it doesn't matter very much - we can read csi of all logs and filter out everything we don't need. But for clusters with small records reading everything would be very expensive, so the AllLogsIterator needs to be able to jump over the logs we don't need, and ideally also over epoch ranges we don't need. This diff implements that.

There are different ways to do it. The ones I considered:
 (1) Make ReadFilter able to reject a record and skip to the next log, kind of like compaction filter but without using directory. But then we'd still have to read the first record of every log in every partition.
 (2) Make AllLogsIterator look at logsdb directory and ask ReadFilter about each directory entry. That's what this diff does.

For how to access the directory, I went with what seems like the simplest solution: when creating the iterator, copy the entire directory from logsdb in-memory view into the iterator. I kept the support for not doing that and reading everything.

This ended up more complicated than I wanted, because AllLogsIterator and CSIWrapper need to somehow work together to skip ranges of logs inside each partition. The way I implemented it is AllLogsIterator tells CSIWrapper to only read until the end of current log (by assigning ReadStats::stop_reading_after, formerly known as stop_reading_after_lsn), and the end of log is hit AllLogsIterator moves on to the next directory entry. (When we're not filtering out much, this may be less efficient than nothing because of the added seeks, but I don't expect the difference to be noticeable.)

Another annoying complication is the special handling of unpartitioned CF.

(This diff turned out pretty convoluted. I'm open to suggestions for how to organize the code better or how to make things less efficient but much simpler. E.g. maybe just do (1) and have CSIWrapper seek itself to the log+epoch requested by filter? I'd sort of prefer to keep this code though, just because it's already written.)

Reviewed By: sh00s

Differential Revision: D10476390

fbshipit-source-id: 16d23500ce34a90e6eab87e37c5bd24a80fc751c