Implemented finding config and SSL certificate files up from the binary's location in tests
Summary:
Before this diff, you have to run tests from a specific root path, otherwise ssl cert/config files on which the tests rely will not be found.

This diff changes that so that the files are searched for upwards from the binary's location - like we do with `logdeviced` binaries.

Reviewed By: calebmarchent

Differential Revision: D12883487

fbshipit-source-id: a6404f0591b5c53a636e097a00e80fe0b35c6045