Added facility to specify different defaults for docs vs. actual settings default
Summary:
We have a couple of settings which use variable defaults that are decided at compile time (or get overridden afterwards). This causes comparison checks on `settings.md` to fail when the file gets generated in a diferent environment vs. where the test runs.

This diff addresses that by specifying an override for the default value that will be presented in documentation, regardless of the actual default value for the setting.

Reviewed By: calebmarchent

Differential Revision: D12957427

fbshipit-source-id: f07730336689c61d3d593b38437bbaf53df79088