fix server-side overflow bug in dataSize()
Summary:
A fairly common case is not accounted for in the server-side part of
dataSize(), where we sum up the log's data size for partitions matching the
given time range. Specifically, if just a few partitions are used, the binary
search on partitions that have a directory entry for the log might point us to
a partition which is the latest with a starting timestamp less than the lower
time bound of the request, but if some partitions after it are unused, i.e. the
next directory entry is not for the very next partition, the partition might
end before the lower time bound as well.

Reviewed By: al13n321

Differential Revision: D12953698

fbshipit-source-id: 1526815679ab67fb7f15b9769897327550c56c56