Timers redesign, implemented WheelTimer class which will dispatchtimers through folly::HHWheelTimer.
Summary:
This diff separates timers scheduling from worker's event base loops and moves the scheduling to the separate threads.  This diff is a part of a worker redesign project which will allow us to move tasks from thread to thread and we cannot have timers attached to the specific thread.

Overview:

We move everything to a separate thread. That thread will contain folly::HHWheelTimer and consequently folly::EventBase for now, because HHWeelTimer requires TimeoutManager interface to attach, and EventBase implements it.

To address the problem of async creation/cancellation problem we divided timer and it consists of 2 parts: scheduling part (WheelTimer.cpp aka wrapper around HHWheelTimer and a thread with EventBase) and executing part (Timer.cpp which provides interface similar to logdevice::LibeventTimer with compatibility middleware). The executing part will be executing all callbacks on a required thread and if the executing part is destroyed or the timer is cancelled (a destruction == a local cancellation in my implementation) it will not be propagated to a required thread and called.

Reviewed By: mcrnic

Differential Revision: D10488678

fbshipit-source-id: 682d974ba5313060235ffe884101f255cbe53739