Fix assertion failure in Socket when forwarding STORE messages
Summary:
It expected that no messages would be enqueued to an incoming socket until an outgoing ACK is passed to TCP. That's not quite valid: Socket starts serializing and enqueueing messages as soon as it receives HELLO (see the `flushSerializeQueue()` call), right after enqueuing an ACK. (It's almost valid though: normally no one would want to send messages into an incoming connection that didn't have a chance to receive any messages yet. But stale ClientID-s in StoreChainLink, combined with ClientID collisions, break this assumption.)

This diff fixes it and adds some more checks along the way. It also adds extra checks when sending STORED messages, to avoid unnecessary and confusing error messages.

Reviewed By: gdrane

Differential Revision: D12969718

fbshipit-source-id: 230fafc757759a53ff68a5bca3099b245bdfd138