Adding an option to trim the data log to first record (or tail lsn if there are no records) before trimming the metadata log
Summary:
The current metadata log trimming behaviour finds the trim point for a data log and trims to it. However, for logs that are empty this works badly - if there have been sequencer reactivations/metadata re-provisions afterwards, the trim point of the data log hasn't moved, and thus the metadata for these empty epochs isn't trimmed.

This diff extracts the data log trim point fetching / trimming logic into `TrimDataLogRequest`, which gets the tail LSN of the log, then it reads from LSN_OLDEST to the tail.

This request can work in 2 modes depending on the `do_trim` argument passed to its constructor:
1) if `do_trim` is false, it just fetches the trim point (matches the previous behaviour).
2) if `do_trim` is true, it trims the data log to either the tail LSN or (first_record_lsn-1), if any records are found. Then it returns the LSN of the new trim point to `TrimMetaDataLogRequest`, which trims the metadata log appropriately.

The default is `do_trim == false`, as trimming delta logs of RSMs breaks them.

Reviewed By: runemaster

Differential Revision: D12905062

fbshipit-source-id: c661142557ae0395be4f8e289cd2794819030025