Ensuring ClientID gets retired only after all users that need it to exist go away.
Summary:
In current worker model socket is bound to a thread and its user is also bound
to the same thread. Because of this if someone wants to send Message they can
use clientID and it will either send the message or close the socket. If some
state machine or user wants to know that the socket was closed, it can register a
callback which will be called when the socket closes. Once the socket is closed
and it gets reclaimed the clientID for the socket can be reused.

With new worker model, where network threads and state machines work on
different threads. Using ClientID to send Message because of queuing delays
between threads could lead send to an incorrect destination. There needs to be
a way for clientIDs to stick around till all the users get to know the socket
is closed and they drop the clientID or clear the context. This diff attempts
to solve the problem.

Here we create a SocketRef which is a cacheable copy of the socket. It
internally contains the ref counted shared_ptr which provides access to the
socket. This way we avoid the socket itself to be a shared_ptr( we can do this
as well).

Whenever a user wants to send message, using the socketref it gets to the
socket if it's still open and sendmessage. If the socket is closed it gets a
nullptr at this point the user needs to drop the cached socket ref so that it
can be cleaned.

In the new worker model, instead of exposing getSocket we can allow clients
to sendMessage directly using socketref so that it does the necessary checks
and the enqueue the message to the network threadpool to send it over the
actual socket.

Reviewed By: mcrnic

Differential Revision: D12942743

fbshipit-source-id: 0e24441e775a5420bb68daf65082dac5ea18e8e2