EventLoopTaskQueue to post folly functions to the EventLoop
Summary:
Changing core of request pump to have UnboundedQueue instead of logdevice/MPSC
queue. Main motivation of this was so that we can post arbitrary folly
functions to the EventLoop which can be executed on the worker instead of
limiting ourselves to Request. This will help with the new worker model.

Based on test we found that there is a small regression with this approach because of allocs and deallocs on the enqueue and dequeue path. The regression is in 20-30 ns but because our average latencies are in milliseconds this small regression does not show up on cluster or on any node operation. Hence it should be ok to go forward with.

Reviewed By: mcrnic

Differential Revision: D10499437

fbshipit-source-id: e91b58f7d965461e354301dc9c0dec9aec6afeb9