ClientFactory
Summary: Currently the `Client` constructor requires you to specify a bunch of arguments, most of which the users don't actually care about. This fixes the issue by changing the API to a factory, where everyone can specify the arguments they care about and rely on sensible defaults if they don't. Most of `ClientFactory::create()` is just moved from `Client::create()`.

Reviewed By: AhmedSoliman

Differential Revision: D13005118

fbshipit-source-id: 814df92d4addb7f40709a786c63697c801ab75b5