Forwarding new memtable allocated notification and tracking new memtable notifications.
Summary:
Memtable factory gets notification of column family for which memtable is
getting allocated. This newly allocated memtable will be used to save new
writes to column family. This notification is important because:
1. This is when unique flush token is assigned to the memtable.
2. This notification is used to note down the event that memtable was allocated
for the column family.
3. The allocated flush token uniquely identifies current active memtable for the
column family.

This information is used as follows:
1. Writes mark the corresponding dependent metadata memtable for this active
memtable.
2. Writes will return flush token value of active memtable to whosoever is
interested in making sure that data written in this memtable is persisted
successfully. Once this and dependent memtables are flushed, the flush token
value will be broadcasted to everyone, letting them know that data was
persisted.
3. Flush callback will check whether any dependent metadata memtable exists for
this memtable and if it exists all the dependencies are flushed along with
this memtable.

Reviewed By: scsiguy

Differential Revision: D9659255

fbshipit-source-id: e9e3e3b43125f4a195ae1622ea63a0810fe9c7fa