fix undefined symbols in `:Configuration` and rename [3/n]
Summary:
This diff resolves the need for `undefined_symbols` in
`logdevice/common:Confguration`, and renames it to conform to the lowercase
target names.

The most semantic change is in `PayloadHolder`. Instead of tracking worker ids
and types, PH now tracks the EventLoop an evbuffer should be associated with.
Which makes sense since evbuffer is tied to an libevent eventbase instead of a
worker.

Reviewed By: mcrnic

Differential Revision: D13089374

fbshipit-source-id: ac1543c54df8a0a1c7773c61c3f774074b315bf2