Fix rejecting recovery stores with NOSPC
Summary: Stores from recovery are supposed to be allowed when storage node's shard is in NOSPC state, see StoreStorageTask::allowIfStoreIsNotAcceptingWrites(). But before the code that implements this special case was hit, there was another check for NOSPC that didn't exempt recovery stores, causing recoveries to get stuck when some nodes are out of space. This diff fixes it.

Differential Revision: D10386877

fbshipit-source-id: a6c015ec314615ab9f299f8b75544174ca870357