More informative last_batch_status in 'info readers'
Summary:
I was trying to investigate this stuck server read stream: P60336107 , and this diff would have been very useful for that.

ServerReadStream::last_batch_status_ is only assigned in a minority of code paths in CatchupOneStream (when storage task or nonblocking read is done) and doesn't reflect the most unusual conditions, like waiting for LNG to be recovered. This diff makes last_batch_status_ into a string instead of a Status, and always assigns it before returning from CatchupOneStream.

Differential Revision: D13080112

fbshipit-source-id: 15dadd41aa4a36b82a34268e0bdf1f3525ace8d3