Fix something about seals and rebuilding
Summary: My previous fix used a fake sequencer node id for rebuilding stores for metadata logs that are not in config. This didn't work out because purging would get stuck if the log is not in config. This diff uses an invalid node id instead, and makes storage nodes skip purging in this case.

Reviewed By: pvjykumar

Differential Revision: D13147143

fbshipit-source-id: 78378ba99e55f3d6c72b79732b582cc37f62df8e