Fixing sequencer activation with boycotting
Summary:
Check the summary of the previous diff in the stack.

This diff forces the secondary sequencer to take over in case its preemptor (the primary) is boycotted.

On a side note: I hate that the logic for when to take over is distributed across multiple places in the codebase.

Differential Revision: D13152887

fbshipit-source-id: 0322f6b4fe11435e9690c41cb711517c6d89ad2b