Implementing the "Multiplicative Increase Additive Decrease Boycott Adaptive Duration" algorithm
Summary:
Implementing the Multiplicative Increase Additive Decrease algorithm. In T36669342, we decided that we won't actually apply the positive feedback to the value to avoid conflicts gossiping it. Instead, each node will independently calculate the current effective duration for a node, given the current boycott duration and the time of the last negative feedback. So basically the duration for a node decreases over time until it gets punished again.

Ps: Is `MultiplicativeIncreaseAdditiveDecreaseBoycottAdaptiveDuration` the biggest class name in the codebase? :D

Reviewed By: adrienconrath

Differential Revision: D13042864

fbshipit-source-id: d22251f6052a1c2be001a25f15e27ed59888e5c4