Debian Packaging for LogDevice
Going to try "3.0 (native)" format initially, if I remember correctly
this means we don't need the upstream tarball.

Changed string literal to define; to avoid error in Chunkbuilding
because debuild enables -Werror=format-security in build config.

debhelper supports cmake based builds, by choosing to place debian tree
in the logdevice subdir the majority of actions work without
intervention. At the time of writing, a couple of the unit tests failed
during debuild, these will be investigated and re-enabled in a future
diff.

In this initial version, build of Python extensions is disabled, it is
planned that these become separate packages, built via distutils which
should then integrate well with debhelper.