Mark dependent metadata memtable during writes.
Summary:
When a partition is written we update the corresponding directory entry in the
metadata column family. This creates a dependency between the two
memtables. At the time of flush, its mandatory to flush the dependent metadata
memtable before data column family memtable. The dependencies between memtables
is established after write to the partition memtable. This diff adds code to mark the dependent memtables.

Reviewed By: al13n321

Differential Revision: D9695039

fbshipit-source-id: fc0fa93036ed39f0626196f15410c70307b71e11