Configuration: Reject the config w/ nodes that 1) does not have storage role and 2) has !=1 generation number
Summary:
This is part of the ongoing effort for getting rid of
generation in logdevice and moving to the new NodesConfiguration. The
first step is to make `generation` a storage only attribute. In this
patch logdevice will reject configuration that has node that does not
have storage role and has >1 generation number.

We use generation number in IntegrationTestUtils to distinguish
replacement lifecyles and determine node data directory. To address
that problem, a separated `replacement_counter` is used in
`IntegrationTestUtils` to keep track of replacement events and determine node directories.

Reviewed By: Zepheus

Differential Revision: D13164150

fbshipit-source-id: e024eec302cdd0d4c3aef2e09ba7dbd50dce0fc8