Implementing the GraylistingTracker class
Summary:
Implement the graylisting tracker class where all the graylisting logic will be. Currently the class is not integrated with anything. The integration will happen in later diffs.

To make `WorkerTimeoutStats` testable, I modified it a bit to pass the time as a param instead of using the clock internally.

For the first version:
- New graylists will eject older ones after reaching the threshold, this is probably not a good idea, but I'll change this behavior in later diffs in the stack anyways.
- I reused some of the flags of the old graylisting given that they shouldn't be both turned on at the same time. The flag to turn this graylisting off, will be introduced during the integration.
- The graylisting duration is fixed for now, later on, it will be adaptive.

Reviewed By: tau0

Differential Revision: D13179211

fbshipit-source-id: c85892377ab1132232c68e85480ee1c137b66b1e